requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
requests>=2.32.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
tavily-python>=0.5.0
qdrant-client[fastembed]>=1.12.0
mem0ai>=0.1.60
